+++ vars/terraform.groovy
@@ -226,6 +226,12 @@ void init(config) {
   if (config.backend == false) {
     cmd += ' -backend=false'
   }
+  if (config.migrateState == true) {
+    cmd += ' -migrate-state'
+  }
+  if (config.forceCopy == true) {
+    cmd += ' -force-copy'
+  }
   if (config.backendConfig) {
     assert (config.backendConfig instanceof List) : 'Parameter backendConfig must be a list of strings.'